Refactor HTTP request
