+++ http_client/src/main/kotlin/model/HttpClientRequest.kt
@@ -14,6 +14,7 @@ data class HttpClientRequest(
     override val port: Int = 80,
     override val path: String = "",
     override val queryString: String = "",
+    override val queryParameters: MultiMap<String, String> = multiMapOf(),
     override val headers: MultiMap<String, String> = multiMapOf(),
     override val body: Any = "",
     override val parts: List<HttpPart> = emptyList(),
